+++ src/components/Popover/DropdownPopover.js
@@ -1,6 +1,7 @@
 import * as React from 'react';
 import { css, cx } from 'emotion';
 import { colors, zIndex, INTERACTIVE_ELEMENT_SELECTOR } from '../../variables';
+import { isFocusInsideRef } from '../../utils/isFocusInside';
 
 const styles = {
   popover: css`
@@ -53,31 +54,24 @@ type DropdownPopoverProps = {
 }
 
 export class DropdownPopover extends React.Component<DropdownPopoverProps> {
-  isFocusInside() {
-    const focused = document.activeElement;
-    const { innerRef } = this.props;
-
-    return innerRef
-      && innerRef.current
-      && (
-        innerRef.current.contains(focused)
-        || innerRef.current === focused
-      );
-  }
-
   focusFirstInteractiveElement = () => {
     const { innerRef } = this.props;
     innerRef.current && focusFirstInteractiveElement(innerRef.current);
   }
 
+  focusPopover = () => {
+    const { innerRef } = this.props;
+    innerRef.current && innerRef.current.focus();
+  }
+
   handleFocusOutside = (evt: FocusEvent) => {
-    if (!this.isFocusInside()) {
+    if (!isFocusInsideRef(this.props.innerRef)) {
       this.focusFirstInteractiveElement();
     }
   }
 
   componentDidMount() {
-    this.focusFirstInteractiveElement();
+    this.focusPopover();
     document.addEventListener('focus', this.handleFocusOutside, true);
   }
 
@@ -86,8 +80,8 @@ export class DropdownPopover extends React.Component<DropdownPopoverProps> {
   }
 
   componentDidUpdate() {
-    if (!this.isFocusInside()) {
-      this.focusFirstInteractiveElement();
+    if (!isFocusInsideRef(this.props.innerRef)) {
+      this.focusPopover();
     }
   }