+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Fixed PageLayout heading shrinking bug.
 
+- Fixed horizontal scrolling bug with copy button in CodeBlock.
+
 ## [0.32.0] - 2020-12-30
 
 - Fixed disabled state in Input and TextArea components.