+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.34.0] - 2021-01-27
 
 - Changed Popover API: prop `popoverContent` accepts function,
   which returns `React.Node` instead of plain `React.Node`.
@@ -21,7 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Added SectionPreloader component.
 
-- Fixed copy button background in CodeBlock.
+- Fixed copy button background and position in CodeBlock.
 
 ## [0.33.0] - 2021-01-13